chore: move archive cache to p2p package
