+++ yarn-project/p2p/src/msg_validators/tx_validator/index.ts
@@ -8,3 +8,4 @@ export * from './gas_validator.js';
 export * from './phases_validator.js';
 export * from './test_utils.js';
 export * from './allowed_public_setup.js';
+export * from './archive_cache.js';
