+++ yarn-project/p2p/tsconfig.json
@@ -28,9 +28,6 @@
       "path": "../protocol-contracts"
     },
     {
-      "path": "../sequencer-client"
-    },
-    {
       "path": "../simulator"
     },
     {
