+++ yarn-project/sequencer-client/src/tx_validator/tx_validator_factory.ts
@@ -1,6 +1,7 @@
 import { Fr } from '@aztec/foundation/fields';
 import {
   AggregateTxValidator,
+  ArchiveCache,
   BlockHeaderTxValidator,
   DataTxValidator,
   DoubleSpendTxValidator,
@@ -20,7 +21,6 @@ import type {
 import { DatabasePublicStateSource, type PublicStateSource } from '@aztec/stdlib/trees';
 import { GlobalVariables, type Tx, type TxValidator } from '@aztec/stdlib/tx';
 
-import { ArchiveCache } from './archive_cache.js';
 import { NullifierCache } from './nullifier_cache.js';
 
 export function createValidatorForAcceptingTxs(
